pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vishwa"
version = "0.1.0"
description = "Terminal-based agentic coding assistant"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Vishwa Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core LLM APIs
    "anthropic>=0.40.0",
    "openai>=1.54.0",

    # CLI & Terminal UI
    "click>=8.1.7",
    "rich>=13.9.4",
    "prompt-toolkit>=3.0.48",

    # File Operations
    "gitpython>=3.1.43",
    "unidiff>=0.7.5",

    # Data & Config
    "pydantic>=2.10.3",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.1",

    # HTTP requests (for Ollama health checks)
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "types-python-dotenv>=1.0.0",
]

[project.scripts]
vishwa = "vishwa.cli.commands:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vishwa"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "unidiff.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "git.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src/vishwa",
    "--cov-report=term-missing",
    "--cov-report=html",
]